Plotting red/blue lines for z-score differences
Make a visualization for EEG coherence such that if the difference in z-score between 2 electrodes is above 2.3, plot a red line connecting those nodes, and if the difference in z-scores is below -2.3 plot a blue line connecting the nodes. 

Connect red lines between electrodes when z-score differences > 2.3. Connect blue lines between electrodes when z-score differences < -2.3.